+++ server/bin/gold/test-7.13.txt
@@ -10599,7 +10599,7 @@ drwxrwxr-x          - logs/pbench-audit-server
 -rw-rw-r--          0 logs/pbench-audit-server/pbench-audit-server.error
 -rw-rw-r--          0 logs/pbench-audit-server/pbench-audit-server.log
 drwxrwxr-x          - logs/pbench-index
--rw-rw-r--       4633 logs/pbench-index/pbench-index.log
+-rw-rw-r--       4543 logs/pbench-index/pbench-index.log
 drwxrwxr-x          - logs/pbench-report-status
 -rw-rw-r--        372 logs/pbench-report-status/pbench-report-status.log
 drwxrwxr-x          - pbench-move-results-receive
@@ -10671,24 +10671,24 @@ drwxrwxr-x          - tmp
 1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index main -- open tar ball
 1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index main -- generator setup
 1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index main -- begin indexing
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index make_all_actions -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_run_action -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_sosreports -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_sosreports -- end [1 sosreports processed]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_tool_info -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_tool_info -- end [1 tools processed]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_run_action -- end
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_toc_actions -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_toc_actions -- end [26 table-of-contents documents]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_result_data_actions -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_result_data_actions -- end [no result data sources]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_tool_data_actions -- start
-1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index _stdout_procint -- tool-data-indexing: tool proc-interrupts, stdout procint start pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19/1/reference-result/tools-default/b03-h01-1029p/proc-interrupts/proc-interrupts-stdout.txt
-1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index _stdout_procint -- tool-data-indexing: tool proc-interrupts, stdout procint end pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19/1/reference-result/tools-default/b03-h01-1029p/proc-interrupts/proc-interrupts-stdout.txt
-1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index _stdout_keyval -- tool-data-indexing: tool proc-vmstat, stdout keyval start pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19/1/reference-result/tools-default/b03-h01-1029p/proc-vmstat/proc-vmstat-stdout.txt
-1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index _stdout_keyval -- tool-data-indexing: tool proc-vmstat, stdout keyval end pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19/1/reference-result/tools-default/b03-h01-1029p/proc-vmstat/proc-vmstat-stdout.txt
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_tool_data_actions -- end [73393 tool data documents]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index make_all_actions -- end
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer make_all_actions -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_run_action -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_sosreports -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_sosreports -- end [1 sosreports processed]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_tool_info -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_tool_info -- end [1 tools processed]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_run_action -- end
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_toc_actions -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_toc_actions -- end [26 table-of-contents documents]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_result_data_actions -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_result_data_actions -- end [no result data sources]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_tool_data_actions -- start
+1970-01-01T00:00:00.000000 INFO pbench-index.indexer _stdout_procint -- tool-data-indexing: tool proc-interrupts, stdout procint start pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19/1/reference-result/tools-default/b03-h01-1029p/proc-interrupts/proc-interrupts-stdout.txt
+1970-01-01T00:00:00.000000 INFO pbench-index.indexer _stdout_procint -- tool-data-indexing: tool proc-interrupts, stdout procint end pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19/1/reference-result/tools-default/b03-h01-1029p/proc-interrupts/proc-interrupts-stdout.txt
+1970-01-01T00:00:00.000000 INFO pbench-index.indexer _stdout_keyval -- tool-data-indexing: tool proc-vmstat, stdout keyval start pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19/1/reference-result/tools-default/b03-h01-1029p/proc-vmstat/proc-vmstat-stdout.txt
+1970-01-01T00:00:00.000000 INFO pbench-index.indexer _stdout_keyval -- tool-data-indexing: tool proc-vmstat, stdout keyval end pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19/1/reference-result/tools-default/b03-h01-1029p/proc-vmstat/proc-vmstat-stdout.txt
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_tool_data_actions -- end [73393 tool data documents]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer make_all_actions -- end
 1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index main -- done indexing (end ts: 1970-01-01T00:00:00-GMT, duration: 0.00s, successes: 73420, duplicates: 0, failures: 0, retries: 0)
 1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index main -- run-1900-01-01T00:00:00-UTC: b03-h01-1029p/pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19.tar.xz: success
 1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index main -- Finished /var/tmp/pbench-test-server/test-7.13/pbench/archive/fs-version-001/b03-h01-1029p/TO-INDEX/pbench-user-benchmark_mbruzek-test-2_2018.04.10T19.01.19.tar.xz (size 3323572)
