+++ server/bin/gold/test-7.6.txt
@@ -279,7 +279,7 @@ drwxrwxr-x          - logs/pbench-audit-server
 -rw-rw-r--          0 logs/pbench-audit-server/pbench-audit-server.error
 -rw-rw-r--          0 logs/pbench-audit-server/pbench-audit-server.log
 drwxrwxr-x          - logs/pbench-index
--rw-rw-r--       3347 logs/pbench-index/pbench-index.log
+-rw-rw-r--       3277 logs/pbench-index/pbench-index.log
 drwxrwxr-x          - logs/pbench-report-status
 -rw-rw-r--        372 logs/pbench-report-status/pbench-report-status.log
 drwxrwxr-x          - pbench-move-results-receive
@@ -351,20 +351,20 @@ drwxrwxr-x          - tmp
 1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index main -- open tar ball
 1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index main -- generator setup
 1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index main -- begin indexing
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index make_all_actions -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_run_action -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_sosreports -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_sosreports -- end [1 sosreports processed]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_tool_info -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_tool_info -- end [1 tools processed]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_run_action -- end
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_toc_actions -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_toc_actions -- end [4 table-of-contents documents]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_result_data_actions -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_result_data_actions -- end [no result data sources]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_tool_data_actions -- start
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index mk_tool_data_actions -- end [0 tool data documents]
-1970-01-01T00:00:00.000000 DEBUG pbench-index.pbench-index make_all_actions -- end
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer make_all_actions -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_run_action -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_sosreports -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_sosreports -- end [1 sosreports processed]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_tool_info -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_tool_info -- end [1 tools processed]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_run_action -- end
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_toc_actions -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_toc_actions -- end [4 table-of-contents documents]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_result_data_actions -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_result_data_actions -- end [no result data sources]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_tool_data_actions -- start
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer mk_tool_data_actions -- end [0 tool data documents]
+1970-01-01T00:00:00.000000 DEBUG pbench-index.indexer make_all_actions -- end
 1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index main -- done indexing (end ts: 1970-01-01T00:00:00-GMT, duration: 0.00s, successes: 5, duplicates: 0, failures: 0, retries: 0)
 1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index main -- run-1900-01-01T00:00:00-UTC: alphaville/test-7.6.tar.xz: success
 1970-01-01T00:00:00.000000 INFO pbench-index.pbench-index main -- Finished /var/tmp/pbench-test-server/test-7.6/pbench/archive/fs-version-001/alphaville/TO-INDEX/test-7.6.tar.xz (size 1492)
